Remove getAllMetadataAndSourcesFromSolcJson outside the local compiler file
