+++ services/server/src/server/controllers/verification/solc-json/session/solc-json.session.handlers.ts
@@ -11,14 +11,14 @@ import {
   PathContent,
 } from "@ethereum-sourcify/lib-sourcify";
 import { BadRequestError } from "../../../../../common/errors";
-import { getAllMetadataAndSourcesFromSolcJson } from "../../../../services/compiler/local/solidityCompiler";
+import { Services } from "../../../../services/services";
 
 export async function addInputSolcJsonEndpoint(req: Request, res: Response) {
   const inputFiles = extractFiles(req, true);
   if (!inputFiles) throw new BadRequestError("No files found");
 
   const solc = req.app.get("solc") as ISolidityCompiler;
-
+  const { verification } = req.app.get("services") as Services;
   const compilerVersion = req.body.compilerVersion;
 
   for (const inputFile of inputFiles) {
@@ -31,11 +31,12 @@ export async function addInputSolcJsonEndpoint(req: Request, res: Response) {
       );
     }
 
-    const metadataAndSources = await getAllMetadataAndSourcesFromSolcJson(
-      solc,
-      solcJson,
-      compilerVersion,
-    );
+    const metadataAndSources =
+      await verification.getAllMetadataAndSourcesFromSolcJson(
+        solc,
+        solcJson,
+        compilerVersion,
+      );
     const metadataAndSourcesPathContents: PathContent[] =
       metadataAndSources.map((pb) => {
         return { path: pb.path, content: pb.buffer.toString(FILE_ENCODING) };
